// src/types.ts
export enum ToolState {
  DASHBOARD = 'dashboard',
  MERGE = 'merge',
  SPLIT = 'split',
  COMPRESS = 'compress',
  EDIT = 'edit',
  ORGANIZE = 'organize',
  EXTRACT = 'extract',
  WATERMARK = 'watermark',
  BACKGROUND = 'background',
  HEADER_FOOTER = 'header_footer',
  CROP = 'crop',
  PAGE_BOXES = 'page_boxes',
  PROPERTIES = 'properties',
  INSERT_BLANK = 'insert_blank',
  INSERT_FILE = 'insert_file',
  REPLACE_PAGES = 'replace_pages',
  BOOKMARK = 'bookmark',
  ATTACHMENT = 'attachment',
  ROTATE_LEFT = 'rotate_left',
  ROTATE_RIGHT = 'rotate_right',
  DELETE_PAGE = 'delete_page',
  UPDATE_WATERMARK = 'update_watermark',
  REMOVE_WATERMARK = 'remove_watermark',
  UPDATE_BACKGROUND = 'update_background',
  REMOVE_BACKGROUND = 'remove_background',
  UPDATE_HEADER_FOOTER = 'update_header_footer',
  REMOVE_HEADER_FOOTER = 'remove_header_footer',
  MANAGE_LINKS = 'manage_links'
}